backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.76
tensorflow==2.13.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
requests==2.31.0
gunicorn==21.2.0